17 спринт - коммит 1
