+++ app/src/main/java/com/practicum/playlistmaker/di/InteractorModule.kt
@@ -25,7 +25,7 @@ val interactorModule = module {
     }
 
     factory<TracksInteractor> {
-        TracksInteractorImpl(get(), get())
+        TracksInteractorImpl(get())
     }
 }
