+++ app/src/main/java/com/practicum/playlistmaker/di/RepositoryModule.kt
@@ -2,12 +2,8 @@ package com.practicum.playlistmaker.di
 
 import com.practicum.playlistmaker.main.data.MainRepositoryImpl
 import com.practicum.playlistmaker.main.domain.MainRepository
-import com.practicum.playlistmaker.player.data.MediaPlayerRepositoryImpl
-import com.practicum.playlistmaker.player.domain.MediaPlayerRepository
-import com.practicum.playlistmaker.search.data.RepositoryImplForHistoryTrack
-import com.practicum.playlistmaker.search.data.RepositoryImplForTracksList
-import com.practicum.playlistmaker.search.domain.TracksHistoryRepository
-import com.practicum.playlistmaker.search.domain.TracksSearchRepository
+import com.practicum.playlistmaker.search.data.TracksRepositoryImpl
+import com.practicum.playlistmaker.search.domain.TracksRepository
 import com.practicum.playlistmaker.settings.data.SettingsRepositoryImpl
 import com.practicum.playlistmaker.settings.domain.SettingsRepository
 import org.koin.dsl.module
@@ -22,15 +18,7 @@ import org.koin.dsl.module
             SettingsRepositoryImpl(get(), get())
         }
 
-        single<TracksHistoryRepository> {
-            RepositoryImplForHistoryTrack(get())
+        single<TracksRepository> {
+            TracksRepositoryImpl(get(), get())
         }
-
-        single<TracksSearchRepository> {
-            RepositoryImplForTracksList(get())
-        }
-
-//        single<MediaPlayerRepository> {
-//            MediaPlayerRepositoryImpl(get())
-//        }
     }
