+++ app/src/main/java/com/practicum/playlistmaker/player/domain/MediaPlayerInteractor.kt
@@ -5,7 +5,7 @@ interface MediaPlayerInteractor {
     fun getState(): PlayerState
     fun getPlayerReady()
     fun getPlaybackPosition(): Int
-    fun setDataSource(url: String)
+    fun setDataURL(url: String)
     fun play()
     fun pause()
     fun stop()
