+++ app/src/main/java/com/practicum/playlistmaker/player/domain/MediaPlayerInteractorImpl.kt
@@ -14,8 +14,8 @@ class MediaPlayerInteractorImpl(private val mediaPlayer: MediaPlayerRepository) 
         mediaPlayer.getPlayerReady()
     }
 
-    override fun setDataSource(url: String) {
-        mediaPlayer.setDataSource(url)
+    override fun setDataURL(url: String) {
+        mediaPlayer.setDataURL(url)
     }
 
     override fun play() {
