+++ app/src/main/java/com/practicum/playlistmaker/player/domain/MediaPlayerRepository.kt
@@ -4,7 +4,7 @@ interface MediaPlayerRepository {
     fun getPlayerState(): PlayerState
     fun getPlaybackPosition(): Int
     fun getPlayerReady()
-    fun setDataSource(url: String)
+    fun setDataURL(url: String)
     fun play()
     fun pause()
     fun destroy()
