+++ app/src/main/java/com/practicum/playlistmaker/search/domain/TracksRepository.kt
@@ -1,9 +1,11 @@
 package com.practicum.playlistmaker.search.domain
 
 import com.practicum.playlistmaker.search.domain.models.Track
+import com.practicum.playlistmaker.search.domain.models.TracksResponse
 
-interface TracksHistoryRepository {
+interface TracksRepository {
     fun saveToHistory(history: ArrayList<Track>)
     fun loadFromHistory(): ArrayList<Track>
     fun killHistory()
+    fun searchTracks(expression: String): TracksResponse
 }
