--- app/src/main/java/com/practicum/playlistmaker/search/domain/TracksSearchRepository.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.search.domain
-
-import com.practicum.playlistmaker.search.domain.models.TracksResponse
-
-interface TracksSearchRepository {
-    fun searchTracks(expression: String): TracksResponse
-}
