+++ app/src/main/java/com/practicum/playlistmaker/settings/ui/SettingsActivity.kt
@@ -2,7 +2,6 @@ package com.practicum.playlistmaker.settings.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModelProvider
 import com.practicum.playlistmaker.databinding.ActivitySettingsBinding
 import com.practicum.playlistmaker.utils.setDebouncedClickListener
 import com.practicum.playlistmaker.utils.bindGoBackButton
@@ -11,18 +10,12 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 class SettingsActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivitySettingsBinding
-//    private lateinit var viewModelProvider: SettingsViewModel
 private val viewModel: SettingsViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivitySettingsBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-//        viewModelProvider = ViewModelProvider(
-//            this,
-//            SettingsViewModel.getViewModelFactory()
-//        )[SettingsViewModel::class.java]
         bindGoBackButton()
 
         // КНОПКА НОЧНОЙ И ДНЕВНОЙ ТЕМЫ
